Revert "Rename SVG -> Icon"
