--- packages/basehub/react-icon.d.ts
@@ -1,1 +1,0 @@
-export * from "./dist/react-icon";
