--- packages/basehub/react-icon.js
@@ -1,1 +1,0 @@
-export * from "./dist/react-icon";
