+++ packages/basehub/src/react/index.ts
@@ -5,4 +5,4 @@ export {
   type CustomBlocksBase,
   type HandlerProps as RichTextHandlerProps,
 } from "./rich-text/primitive";
-export { SVG, Icon } from "./icon/primitive";
+export { SVG } from "./svg/primitive";
